Add more structure to the vm cli
